requirements.txt
python-dateutil>=2.8.2
requests>=2.31.0
Jinja2>=3.1.2
beautifulsoup4>=4.12.0
lxml>=4.9.0
